[RENAME] 로그인, 뱃지 등록 등 설문조사형 컴포넌트 스타일 Survey로 수정하기 #NADA-60
