+++ src/Auth/components/AuthForm.js
@@ -1,5 +1,5 @@
 // 회원가입 또는 로그인 폼을 보여주기
-import { InputBox, LoginBtn } from '../styles/Auth';
+import { InputBox, LoginBtn } from '../../styles/Survey';
 
 const AuthForm = ({ form, onChange, onSubmit }) => {
   return (
