+++ src/Auth/components/Caution.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { ErrorMessage } from '../styles/Auth';
+import { ErrorMessage } from '../../styles/Survey';
 import { WarningSvg } from '../../icon/WarningSvg';
 
 const CautionBox = styled.div`
