+++ src/Auth/containers/register/FieldForm.js
@@ -1,6 +1,6 @@
 // 회원가입 - 관심분야를 입력받는 컨테이너
 import { authSelector } from '../../modules/redux/auth';
-import { TitleBox, Explain, LoginBtn } from '../../styles/Auth';
+import { TitleBox, Explain, LoginBtn } from '../../../styles/Survey';
 import { useState } from 'react';
 import { fieldData } from '../../../modules/common/AttributeData';
 import { Item } from '../../../components/common/filter/FilterItems';
